+++ src/tasks/minions/fishingActivity.ts
@@ -1,14 +1,14 @@
 import { Task } from 'klasa';
-import { removeBankFromBank } from 'oldschooljs/dist/util';
 
 import { Emoji, Events } from '../../lib/constants';
 import hasArrayOfItemsEquipped from '../../lib/gear/functions/hasArrayOfItemsEquipped';
+import addItemsToBankAndReturn from '../../lib/minions/functions/addItemsToBankAndReturn';
+import addSkillingClueToLoot from '../../lib/minions/functions/addSkillingClueToLoot';
 import { UserSettings } from '../../lib/settings/types/UserSettings';
 import Fishing from '../../lib/skilling/skills/fishing';
 import { SkillsEnum } from '../../lib/skilling/types';
-import { Bank } from '../../lib/types';
 import { FishingActivityTaskOptions } from '../../lib/types/minions';
-import { randFloat, roll } from '../../lib/util';
+import { roll } from '../../lib/util';
 import createReadableItemListFromBank from '../../lib/util/createReadableItemListFromTuple';
 import { handleTripFinish } from '../../lib/util/handleTripFinish';
 import itemID from '../../lib/util/itemID';
@@ -99,36 +99,14 @@ export default class extends Task {
 		if (fish.id === itemID('Raw karambwanji')) {
 			quantity *= 1 + Math.floor(user.skillLevel(SkillsEnum.Fishing) / 5);
 		}
-		const loot = {
+		let loot = {
 			[fish.id]: quantity
 		};
 
 		// Add clue scrolls
 		// https://oldschool.runescape.wiki/w/Clue_bottle
 		if (fish.clueScrollChance) {
-			const clues: Record<number, number> = {
-				[itemID('Clue scroll(easy)')]: 4 / 10,
-				[itemID('Clue scroll(medium)')]: 3 / 10,
-				[itemID('Clue scroll(hard)')]: 2 / 10,
-				[itemID('Clue scroll(elite)')]: 1 / 10,
-				[itemID('Clue scroll(beginner)')]: 1 / 1000
-			};
-			for (let i = 0; i < quantity; i++) {
-				if (
-					roll(
-						Math.floor(
-							fish.clueScrollChance / (100 + user.skillLevel(SkillsEnum.Fishing))
-						)
-					)
-				) {
-					for (const clue of Object.entries(clues)) {
-						if (randFloat(0, 1) <= clue[1]) {
-							loot[Number(clue[0])] = (loot[Number(clue[0])] ?? 0) + 1;
-							break;
-						}
-					}
-				}
-			}
+			loot = addSkillingClueToLoot(user, quantity, fish.clueScrollChance, loot);
 		}
 
 		// Add barbarian fish to loot
@@ -140,6 +118,7 @@ export default class extends Task {
 		}
 
 		str += `\n\nYou received: ___ITEMSRECEIVED___.`;
+
 		if (fish.name === 'Barbarian fishing') {
 			str = `${user}, ${user.minionName} finished fishing ${quantity} ${
 				fish.name
@@ -175,14 +154,11 @@ export default class extends Task {
 		}
 
 		// Show only what is added to the bank, to avoid showing multiple of the same clue scroll
-		const addResult = Object.values(await user.addItemsToBank(loot, true));
-		const bankPrevious = addResult[0].previous as Bank;
-		const bankAfter = addResult[0].next as Bank;
 		str = str.replace(
 			'___ITEMSRECEIVED___',
 			await createReadableItemListFromBank(
 				this.client,
-				removeBankFromBank(bankAfter, bankPrevious)
+				await addItemsToBankAndReturn(user, loot)
 			)
 		);
