+++ src/tasks/minions/woodcuttingActivity.ts
@@ -1,12 +1,12 @@
 import { Task } from 'klasa';
-import { removeBankFromBank } from 'oldschooljs/dist/util';
 
 import { Emoji, Events } from '../../lib/constants';
+import addItemsToBankAndReturn from '../../lib/minions/functions/addItemsToBankAndReturn';
+import addSkillingClueToLoot from '../../lib/minions/functions/addSkillingClueToLoot';
 import Woodcutting from '../../lib/skilling/skills/woodcutting';
 import { SkillsEnum } from '../../lib/skilling/types';
-import { Bank } from '../../lib/types';
 import { WoodcuttingActivityTaskOptions } from '../../lib/types/minions';
-import { randFloat, roll } from '../../lib/util';
+import { roll } from '../../lib/util';
 import createReadableItemListFromBank from '../../lib/util/createReadableItemListFromTuple';
 import { handleTripFinish } from '../../lib/util/handleTripFinish';
 import itemID from '../../lib/util/itemID';
@@ -34,39 +34,16 @@ export default class extends Task {
 			str += `\n\n${user.minionName}'s Woodcutting level is now ${newLevel}!`;
 		}
 
-		const loot = {
+		let loot = {
 			[Log.id]: quantity
 		};
 
 		// Add clue scrolls
-		// https://oldschool.runescape.wiki/w/Clue_bottle
 		if (Log.clueScrollChance) {
-			const clues: Record<number, number> = {
-				[itemID('Clue scroll(easy)')]: 4 / 10,
-				[itemID('Clue scroll(medium)')]: 3 / 10,
-				[itemID('Clue scroll(hard)')]: 2 / 10,
-				[itemID('Clue scroll(elite)')]: 1 / 10,
-				[itemID('Clue scroll(beginner)')]: 1 / 1000
-			};
-			for (let i = 0; i < quantity; i++) {
-				if (
-					roll(
-						Math.floor(
-							Log.clueScrollChance / (100 + user.skillLevel(SkillsEnum.Woodcutting))
-						)
-					)
-				) {
-					for (const clue of Object.entries(clues)) {
-						if (randFloat(0, 1) <= clue[1]) {
-							loot[Number(clue[0])] = (loot[Number(clue[0])] ?? 0) + 1;
-							break;
-						}
-					}
-				}
-			}
+			loot = addSkillingClueToLoot(user, quantity, Log.clueScrollChance, loot);
 		}
 
-		// roll for pet
+		// Roll for pet
 		if (
 			Log.petChance &&
 			roll((Log.petChance - user.skillLevel(SkillsEnum.Woodcutting) * 25) / quantity)
@@ -79,19 +56,11 @@ export default class extends Task {
 			);
 		}
 
-		str += `\n\nYou received: ___ITEMSRECEIVED___.`;
-
 		// Show only what is added to the bank, to avoid showing multiple of the same clue scroll
-		const addResult = Object.values(await user.addItemsToBank(loot, true));
-		const bankPrevious = addResult[0].previous as Bank;
-		const bankAfter = addResult[0].next as Bank;
-		str = str.replace(
-			'___ITEMSRECEIVED___',
-			await createReadableItemListFromBank(
-				this.client,
-				removeBankFromBank(bankAfter, bankPrevious)
-			)
-		);
+		str += `\n\nYou received: ${await createReadableItemListFromBank(
+			this.client,
+			await addItemsToBankAndReturn(user, loot)
+		)}.`;
 
 		handleTripFinish(this.client, user, channelID, str, res => {
 			user.log(`continued trip of ${quantity}x ${Log.name}[${Log.id}]`);
